actual paid price from int to float

+++ back-end/src/main/java/com/techtitans/backend/entity/PaymentEntity.java
@@ -8,7 +8,7 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @NoArgsConstructor
 public class PaymentEntity {
-    private int actualPaidPrice;
+    private float actualPaidPrice;
     private int backerId;
     private int benefitId;
 }
